Adding validations

+++ APIRestAzureFuncCosmos/Presentation/TaskFunction.cs
@@ -7,18 +7,22 @@ using Microsoft.AspNetCore.Http;
 using Newtonsoft.Json;
 using Application.DTOs;
 using Application.Interfaces;
-using Presentation.Enums;
 using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
 using Microsoft.OpenApi.Models;
 using System.Net;
 using System.Collections.Generic;
+using Shared.Enums;
+using Shared.Validations;
+using FluentValidation;
+using System.Linq;
 
 namespace Presentation;
 
-public class TaskFunction(ITaskService taskService)
+public class TaskFunction(ITaskService taskService, IValidator<TaskDTO> createValidator)
 {
     const string ROUTE_NAME = "task";
     private readonly ITaskService _taskService = taskService;
+    private readonly IValidator<TaskDTO> _createValidator = createValidator;
 
     /// <summary>
     /// Get all tasks.
@@ -59,11 +63,17 @@ public class TaskFunction(ITaskService taskService)
     [OpenApiOperation(operationId: nameof(GetTaskById), tags: [ROUTE_NAME])]
     [OpenApiParameter(name: nameof(id), In = ParameterLocation.Path, Required = true, Type = typeof(string), Description = "Task's id")]
     [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: RestUtilityConsts.APPJSON, bodyType: typeof(TaskDTO), Description = "Get a task by id")]
+    [OpenApiResponseWithBody(statusCode: HttpStatusCode.BadRequest, contentType: RestUtilityConsts.APPJSON, bodyType: typeof(ErrorResponse), Description = "Model for errors")]
     [OpenApiResponseWithoutBody(statusCode: HttpStatusCode.NotFound, Description = "Task was not found")]
     [FunctionName(nameof(GetTaskById))]
     public async Task<IActionResult> GetTaskById(
         [HttpTrigger(AuthorizationLevel.Anonymous, RestUtilityConsts.GET, Route = $"{ROUTE_NAME}/{{id}}")] HttpRequest req, string id)
     {
+        if (string.IsNullOrWhiteSpace(id)) 
+        {
+            return new BadRequestObjectResult(RestUtilityConsts.VALIDATION_ID_NOT_EMPTY);
+        }
+
         var task = await _taskService.GetByIdAsync(id);
         if (task == null)
         {
@@ -102,6 +112,17 @@ public class TaskFunction(ITaskService taskService)
     {
         var requestBody = await new StreamReader(req.Body).ReadToEndAsync();
         var createTaskDto = JsonConvert.DeserializeObject<TaskDTO>(requestBody);
+
+        if (createTaskDto == null)
+        {
+            return new BadRequestObjectResult(new { Error = RestUtilityConsts.VALIDATION_INVALID_JSON_REQUEST });
+        }
+
+        var validationResult = await _createValidator.ValidateAsync(createTaskDto);
+        if (!validationResult.IsValid)
+        {
+            return new BadRequestObjectResult(new { Errors = validationResult.Errors.Select(e => e.ErrorMessage) });
+        }
 
         var createdTask = await _taskService.CreateAsync(createTaskDto);
 
@@ -144,12 +165,19 @@ public class TaskFunction(ITaskService taskService)
         [HttpTrigger(AuthorizationLevel.Function, RestUtilityConsts.PUT, Route = $"{ROUTE_NAME}")] HttpRequest req)
     {
         var requestBody = await new StreamReader(req.Body).ReadToEndAsync();
-        var taskToUpdate = JsonConvert.DeserializeObject<TaskDTO>(requestBody);
+        var updateTaskDto = JsonConvert.DeserializeObject<TaskDTO>(requestBody);
 
-        if (taskToUpdate == null || string.IsNullOrWhiteSpace(taskToUpdate.Id))
-            return new BadRequestObjectResult("Invalid task data");
+        if (updateTaskDto == null)
+        {
+            return new BadRequestObjectResult(new { Error = RestUtilityConsts.VALIDATION_INVALID_JSON_REQUEST });
+        }
 
-        var updatedTask = await _taskService.UpdateAsync(taskToUpdate);
+        if (string.IsNullOrWhiteSpace(updateTaskDto.Id))
+        {
+            return new BadRequestObjectResult(RestUtilityConsts.VALIDATION_ID_NOT_EMPTY);
+        }
+
+        var updatedTask = await _taskService.UpdateAsync(updateTaskDto);
         if (updatedTask == null)
         { 
             return new NotFoundResult();
@@ -176,10 +204,16 @@ public class TaskFunction(ITaskService taskService)
     [OpenApiOperation(operationId: nameof(DeleteTask), tags: [ROUTE_NAME])]
     [OpenApiParameter(name: nameof(req), In = ParameterLocation.Path, Required = true, Type = typeof(string), Description = "The id of the task to be delete")]
     [OpenApiResponseWithoutBody(statusCode: HttpStatusCode.NoContent, Description = "Deletes a task")]
+    [OpenApiResponseWithBody(statusCode: HttpStatusCode.BadRequest, contentType: RestUtilityConsts.APPJSON, bodyType: typeof(ErrorResponse), Description = "Model for errors")]
     [OpenApiResponseWithoutBody(statusCode: HttpStatusCode.NotFound, Description = "Task was not found")]
     public async Task<IActionResult> DeleteTask(
         [HttpTrigger(AuthorizationLevel.Function, RestUtilityConsts.DELETE, Route = $"{ROUTE_NAME}/{{id}}")] HttpRequest req, string id)
-    {   
+    {
+        if (string.IsNullOrWhiteSpace(id))
+        {
+            return new BadRequestObjectResult(RestUtilityConsts.VALIDATION_ID_NOT_EMPTY);
+        }
+
         var deleted = await _taskService.DeleteAsync(id);
         if (!deleted)
         {
